refactor: move products to module folder
