+++ src/modules/products/entities/color.entity.ts
