+++ src/modules/products/entities/colors.enum.ts
