+++ src/modules/products/entities/styles.enum.ts
