+++ src/modules/products/products.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { Color } from 'src/products/entities/color.entity';
-import { Product } from 'src/products/entities/product.entity';
-import { ProductsController } from 'src/products/products.controller';
-import { ProductsService } from 'src/products/products.service';
+import { Color } from 'src/modules/products/entities/color.entity';
+import { Product } from 'src/modules/products/entities/product.entity';
+import { ProductsController } from 'src/modules/products/products.controller';
+import { ProductsService } from 'src/modules/products/products.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Product, Color])],
